UserBot_Backup_20250707_003942/pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "asyncio-mqtt>=0.16.2",
    "flask-cors>=6.0.1",
    "flask>=3.1.1",
    "openai>=1.93.0",
    "psutil>=6.1.0",
    "py-tgcalls>=2.1.0",
    "pytz>=2025.2",
    "requests>=2.32.4",
    "telethon>=1.40.0",
    "youtube-dl>=2021.12.17",
    "yt-dlp>=2025.6.30",
    "python-dotenv>=1.1.1",
    "ffmpeg-python>=0.2.0",
    "asyncio-throttle>=1.0.2",
]
